# api/openapi/kapsora-v1.yaml
openapi: 3.1.0
info:
  title: KAPSORA Platform API
  version: 1.0.0-alpha.1
  summary: Kurumsal hak, fayda ve hizmet orkestrasyonu için başlangıç API sözleşmesi
  description: |
    Bu sözleşme ilk sprint ve çekirdek MVP için contract-first geliştirme başlangıç noktasıdır.
    Kullanıcılar KAPSORA'nın kendi hesaplarıyla giriş yapar (ADR-022); tarayıcı yalnız opak
    oturum çerezi taşır. Makineden makineye entegrasyonlar OAuth 2.0 client credentials ve
    mTLS ile çalışır. Tüm tenant kapsamlı çağrılar X-Tenant-ID
    başlığı taşır. Durum değiştiren POST komutlarında Idempotency-Key zorunludur; sorgu
    niteliğindeki POST'larda (eligibility check, pricing quote) opsiyoneldir.
    İstemci X-Request-ID başlığında UUID gönderebilir; sunucu her yanıtta X-Request-ID döner.
    Çerez oturumuyla yapılan durum değiştiren çağrılar X-CSRF-Token başlığı taşır.
  contact:
    name: KAPSORA Product Engineering
servers:
  - url: https://api.staging.kapsora.example
    description: Staging
  - url: https://api.kapsora.example
    description: Production
security:
  - sessionCookie: []
  - oauth2: [kapsora.api]
...
paths:
...
  /api/v1/contract-versions/{contractVersionId}/price-lists:
    put:
      tags: [Contracts]
      operationId: putPriceLists
      description: |
        Replaces the whole price list set of a DRAFT version, for the same reason provider
        capabilities are replaced as a set: the meaningful unit is the price sheet, and a
        set write makes consistency one decision. A list already present under the same
        code keeps its identity and its price items; one absent from the body is deleted
        with its items. Season and weekday mask exist because accommodation prices by date
        and by day of the week, and a health contract simply leaves both null. Any write to
        a version that is not DRAFT answers 409 CONTRACT_VERSION_IMMUTABLE. If-Match
        carries the ETag of the contract version, which the replacement moves on.
      parameters:
        - $ref: '#/components/parameters/TenantHeader'
        - $ref: '#/components/parameters/ContractVersionId'
        - $ref: '#/components/parameters/IfMatch'
        - $ref: '#/components/parameters/CsrfHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReplacePriceListsRequest'
      responses:
        '200':
          description: Price list set replaced
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PriceListList'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '412':
          description: ETag mismatch
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '422':
          $ref: '#/components/responses/ValidationError'
        '428':
          description: If-Match header missing
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
  /api/v1/price-lists/{priceListId}/items:
    get:
      tags: [Contracts]
      operationId: listPriceItems
      description: |
        The price items of one list, oldest first, with keyset paging. Every money value
        crosses the wire as an exact decimal string and never as a JSON number: a tariff
        row is what somebody is invoiced, and a float would silently round it. The ETag is
        the one of the owning price list, which is what a replacement expects in If-Match.
      parameters:
        - $ref: '#/components/parameters/TenantHeader'
        - $ref: '#/components/parameters/PriceListId'
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Price item page
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PriceItemPage'
        '400':
          description: Cursor invalid
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Contracts]
      operationId: putPriceItems
      description: |
        Replaces the whole price item set of one list. Each row prices exactly one of a
        service definition, a service category or a package definition, and may be
        restricted to one location; the specificity ladder of the price selection depends
        on that being true, so a row naming none or more than one is refused. Amounts are
        exact decimal strings. The owning version must be DRAFT, otherwise the answer is
        409 CONTRACT_VERSION_IMMUTABLE. If-Match carries the ETag of the price list.
      parameters:
        - $ref: '#/components/parameters/TenantHeader'
        - $ref: '#/components/parameters/PriceListId'
        - $ref: '#/components/parameters/IfMatch'
        - $ref: '#/components/parameters/CsrfHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReplacePriceItemsRequest'
      responses:
        '200':
          description: Price item set replaced
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PriceItemPage'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '412':
          description: ETag mismatch
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '422':
          $ref: '#/components/responses/ValidationError'
        '428':
          description: If-Match header missing
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
  /api/v1/contract-versions/{contractVersionId}/package-definitions:
    get:
      tags: [Contracts]
      operationId: listPackageDefinitions
      description: |
        The packages of one contract version with their lines. A package is a named bundle
        of service definitions with an included quantity each; a price item pointing at a
        package prices the bundle as a whole rather than its parts. The set is short by
        design, so it is returned whole rather than paged.
      parameters:
        - $ref: '#/components/parameters/TenantHeader'
        - $ref: '#/components/parameters/ContractVersionId'
      responses:
        '200':
          description: Package definitions of the version
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PackageDefinitionList'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Contracts]
      operationId: putPackageDefinitions
      description: |
        Replaces the whole package set of a DRAFT version, lines included. A package
        already present under the same code keeps its identity, so the price items
        pointing at it survive the write; one absent from the body is deleted together
        with the price items that priced it. ANY_OF_N needs minLines and ALL forbids it.
        Any write to a version that is not DRAFT answers 409 CONTRACT_VERSION_IMMUTABLE.
      parameters:
        - $ref: '#/components/parameters/TenantHeader'
        - $ref: '#/components/parameters/ContractVersionId'
        - $ref: '#/components/parameters/IfMatch'
        - $ref: '#/components/parameters/CsrfHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReplacePackageDefinitionsRequest'
      responses:
        '200':
          description: Package set replaced
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PackageDefinitionList'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '412':
          description: ETag mismatch
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '422':
          $ref: '#/components/responses/ValidationError'
        '428':
          description: If-Match header missing
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
  /api/v1/contract-versions/{contractVersionId}/provider-quotas:
    get:
      tags: [Contracts]
      operationId: listProviderQuotas
      description: |
        The capacity the provider granted under this version: how much of a service, at
        which location, over which period. Consuming a quota belongs to authorization, so
        this module only stores and reads them; consumed is shown as it stands.
      parameters:
        - $ref: '#/components/parameters/TenantHeader'
        - $ref: '#/components/parameters/ContractVersionId'
      responses:
        '200':
          description: Provider quotas of the version
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProviderQuotaList'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Contracts]
      operationId: putProviderQuotas
      description: |
        Replaces the whole quota set of a DRAFT version. A quota already present under the
        same scope and period keeps its identity and its consumed counter; two rows with
        the same location, service and period collide, because a scope that means
        "everything" has to be unique too. Capacities are exact decimal strings. Any write
        to a version that is not DRAFT answers 409 CONTRACT_VERSION_IMMUTABLE.
      parameters:
        - $ref: '#/components/parameters/TenantHeader'
        - $ref: '#/components/parameters/ContractVersionId'
        - $ref: '#/components/parameters/IfMatch'
        - $ref: '#/components/parameters/CsrfHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReplaceProviderQuotasRequest'
      responses:
        '200':
          description: Quota set replaced
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProviderQuotaList'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '412':
          description: ETag mismatch
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '422':
          $ref: '#/components/responses/ValidationError'
        '428':
          description: If-Match header missing
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
  /api/v1/contract-versions/{contractVersionId}/payment-term:
    get:
      tags: [Contracts]
      operationId: getPaymentTerm
      description: |
        How and when the provider is paid under this version: the due days, the settlement
        method, the tax behaviour and the rates. A version that has none answers 404, so a
        screen can tell "not agreed yet" apart from "agreed as zero".
      parameters:
        - $ref: '#/components/parameters/TenantHeader'
        - $ref: '#/components/parameters/ContractVersionId'
      responses:
        '200':
          description: Payment term of the version
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentTerm'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Contracts]
      operationId: putPaymentTerm
      description: |
        Writes the single payment term of a DRAFT version, creating it or replacing it in
        place. EXCLUSIVE and INCLUSIVE tax behaviour both require a VAT rate; EXEMPT does
        not. The rates are exact decimal strings. Any write to a version that is not DRAFT
        answers 409 CONTRACT_VERSION_IMMUTABLE.
      parameters:
        - $ref: '#/components/parameters/TenantHeader'
        - $ref: '#/components/parameters/ContractVersionId'
        - $ref: '#/components/parameters/IfMatch'
        - $ref: '#/components/parameters/CsrfHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PutPaymentTermRequest'
      responses:
        '200':
          description: Payment term written
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentTerm'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '412':
          description: ETag mismatch
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '422':
          $ref: '#/components/responses/ValidationError'
        '428':
          description: If-Match header missing
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: __Host-kapsora_session
      description: |
        Opaque session id. The server stores only its SHA-256 hash and the browser never
        holds a token of any kind. Local HTTP development uses the name kapsora_session
        because the __Host- prefix requires HTTPS.
    oauth2:
      type: oauth2
      flows:
        clientCredentials:
          tokenUrl: https://identity.kapsora.example/realms/kapsora/protocol/openid-connect/token
          scopes:
            kapsora.api: Call KAPSORA APIs
  parameters:
    TenantHeader:
      name: X-Tenant-ID
      in: header
      required: true
      description: Selected tenant UUID. It must be one of the actor's active memberships.
      schema:
        type: string
        format: uuid
...
    CsrfHeader:
      name: X-CSRF-Token
      in: header
      required: false
      description: Required when the request is authenticated with the BFF session cookie.
      schema:
        type: string
        minLength: 16
        maxLength: 256
    IfMatch:
      name: If-Match
      in: header
      required: true
      description: Optimistic concurrency token returned as ETag.
      schema:
        type: string
    Cursor:
      name: cursor
      in: query
      schema:
        type: string
      description: Opaque cursor from the previous response.
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
...
    ContractVersionId:
      name: contractVersionId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    PriceListId:
      name: priceListId
      in: path
      required: true
      schema:
        type: string
        format: uuid
  headers:
    ETag:
      description: Strong optimistic concurrency tag based on row_version.
      schema:
        type: string
  responses:
...
    Forbidden:
      description: Authenticated actor lacks permission or scope
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    NotFound:
      description: Resource does not exist or is not visible in the selected tenant
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Conflict:
      description: Business state, duplicate idempotency key, or uniqueness conflict
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    ValidationError:
      description: Request is syntactically valid but violates validation or business preconditions
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
  schemas:
...
    Problem:
      type: object
      additionalProperties: false
      required: [type, title, status, code, traceId]
      properties:
        type:
          type: string
          format: uri-reference
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
          format: uri-reference
        code:
          type: string
          pattern: '^[A-Z][A-Z0-9_.-]{2,79}$'
        traceId:
          type: string
        errors:
          type: array
          items:
            type: object
            additionalProperties: false
            required: [field, code]
            properties:
              field:
                type: string
              code:
                type: string
              message:
                type: string
...
    SettlementMethod:
      type: string
      description: How the provider is paid.
      enum: [BANK_TRANSFER, OFFSET, OTHER]
    TaxBehaviour:
      type: string
      description: Whether the agreed amounts already include VAT, exclude it, or are exempt.
      enum: [EXCLUSIVE, INCLUSIVE, EXEMPT]
...
    PriceListList:
      type: object
      additionalProperties: false
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/PriceList'
    ReplacePriceListsRequest:
      type: object
      additionalProperties: false
      required: [items]
      properties:
        items:
          type: array
          maxItems: 200
          items:
            $ref: '#/components/schemas/PriceListInput'
...
    PriceItemPage:
      type: object
      additionalProperties: false
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/PriceItem'
        nextCursor:
          type: [string, 'null']
    ReplacePriceItemsRequest:
      type: object
      additionalProperties: false
      required: [items]
      properties:
        items:
          type: array
          maxItems: 2000
          items:
            $ref: '#/components/schemas/PriceItemInput'
...
    PackageDefinitionList:
      type: object
      additionalProperties: false
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/PackageDefinition'
    ReplacePackageDefinitionsRequest:
      type: object
      additionalProperties: false
      required: [items]
      properties:
        items:
          type: array
          maxItems: 200
          items:
            $ref: '#/components/schemas/PackageDefinitionInput'
...
    ProviderQuotaList:
      type: object
      additionalProperties: false
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ProviderQuota'
    ReplaceProviderQuotasRequest:
      type: object
      additionalProperties: false
      required: [items]
      properties:
        items:
          type: array
          maxItems: 500
          items:
            $ref: '#/components/schemas/ProviderQuotaInput'
    PaymentTerm:
      type: object
      additionalProperties: false
      required: [id, contractVersionId, dueDays, settlementMethod, taxBehaviour, rowVersion]
      properties:
        id:
          type: string
          format: uuid
        contractVersionId:
          type: string
          format: uuid
        dueDays:
          type: integer
          minimum: 0
          maximum: 365
        settlementMethod:
          $ref: '#/components/schemas/SettlementMethod'
        taxBehaviour:
          $ref: '#/components/schemas/TaxBehaviour'
        vatRate:
          type: [string, 'null']
          description: VAT percentage as an exact decimal string.
        lateFeePercent:
          type: [string, 'null']
          description: Monthly late fee percentage as an exact decimal string.
        rowVersion:
          type: integer
          minimum: 1
...